Create TestCollections.java
Setting up date into the Array List. Data is songs and movies with titles and production year. I need to expand this into an app that reflects a simple media player with GUI that lets user enter an index integer value into a text field and output the title and year to label fields and load an image of the album on the page. I'll need to use java indexOf() method for sure.