Update TestCollections.java
Added random number to simulate user input of an index between 1-20 with help of Math.random, Added a for loop to increment through 20 iterations that do the same thing at each iteration and keep track of the iteration it's on by increment by 1 a value stored in a variable. Inside the for loop I added an in statement to check if the random number that represents an index value for a song or movie choice to see if it's equal to the index value the for loop iteration happens to be on. Each of the number value stored in tracking variable corresponds to a slot number where an element lives. I had to put a get method inside as a parameter for the index of method to get this done because of how parameters and return types have been defined for these methods.