Update TestCollections.java
Changed random number to getting user input from scanner object and println prompt. Will work on exception handling trying to catch input that's not numbers and numbers out of range of 1-20. I'm getting all the images from the songs and albums and sizing them the same. Then i'll store each of the images as a file object in an array list with indexes that match the array list of songs and movies.