Update TestCollections.java
Finally, found out how to combine the validate number between range and check if number is a number while keeping the flow of the program smooth for the user. Basically contained a check for number within range, inside there another contained area locks in the user and continuously asks for a number and when store user input into a variable we wrap the scanner variable in a method that filters it to make sure it's a number before placing it into a variable of integer value.